feat : 7단계 테스트 통과
