--- src/main/java/auth/JwtConfig.java
@@ -1,4 +1,0 @@
-package auth;
-
-public class JwtConfig {
-}
