+++ src/main/java/roomescape/infrastructure/AuthMember.java
@@ -1,4 +1,4 @@
-package auth;
+package roomescape.infrastructure;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
