+++ src/main/java/roomescape/member/MemberController.java
@@ -1,6 +1,5 @@
 package roomescape.member;
 
-import auth.AuthMember;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.http.ResponseEntity;
@@ -8,6 +7,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.infrastructure.AuthMember;
 
 import java.net.URI;
