+++ src/main/java/roomescape/presentation/AuthenticationPrincipalArgumentResolver.java
@@ -1,7 +1,5 @@
 package roomescape.presentation;
 
-import auth.AuthMember;
-import auth.AuthService;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Component;
@@ -9,6 +7,8 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
+import roomescape.infrastructure.AuthMember;
+import roomescape.infrastructure.AuthService;
 import roomescape.member.Member;
 
 @Component
