+++ src/main/java/roomescape/reservation/waiting/WaitingController.java
@@ -1,10 +1,10 @@
 package roomescape.reservation.waiting;
 
-import auth.AuthMember;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.infrastructure.AuthMember;
 import roomescape.member.Member;
 import roomescape.reservation.dto.ReservationRequest;
