+++ src/test/java/roomescape/MissionStepTest.java
@@ -1,34 +1,19 @@
 package roomescape;
 
+import auth.JwtTokenProvider;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
-import roomescape.time.Time;
-import roomescape.time.TimeRepository;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.stereotype.Component;
+import org.springframework.test.annotation.DirtiesContext;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@DataJpaTest
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 public class MissionStepTest {
-
-    @Autowired
-    private TestEntityManager entityManager;
-
-    @Autowired
-    private TimeRepository timeRepository;
-
     @Test
-    void 사단계() {
-        Time time = new Time("10:00");
-        entityManager.persist(time);
-
-        entityManager.flush();
-        entityManager.clear();
-
-        Time persistTime = entityManager.find(Time.class, time.getId());
-
-        assertThat(persistTime).isNotNull();
-        assertThat(persistTime.getValue()).isEqualTo(time.getValue());
+    void 칠단계() {
+        Component componentAnnotation = JwtTokenProvider.class.getAnnotation(Component.class);
+        assertThat(componentAnnotation).isNull();
     }
 }
